# docs/api.yml
openapi: 3.0.3
info:
  version: 0.1.0
  title: Swagger PKPortal - OpenAPI 3.0
  description: Описание API для бекенда портала школы "Покровский Квартал"
...
paths:
...
  /auth/activate/{activationLink}:
    get:
      tags:
        - Auth
      summary: Активация аккаунта
      description: Эндпоинт для активации через активационное письмо email
      operationId: activate
      parameters:
        - name: activationLink
          in: path
          description: Активационный url
          required: true
          schema:
            type: string
      responses:
        200:
          description: ОК
        404:
          description: Пользователь с таким ActivationLink не найден
        500:
          description: Ошибка сервера
...